refactor: reorganize project structure for clarity
Move code into src/ layout and split the monolithic workspace.py into
focused modules: CLI command handlers under commands/ and shared utilities
under utils/. Bundled templates and services move into _static/.

This makes the codebase easier to navigate and maintain without changing
any behavior.
