+++ .gitignore
@@ -1,24 +1,17 @@
-.bare-git-repos/
-.claude/
-ow/services/volumes/
-ow/services/compose.local.yml
-workspaces/*
-.ow/
-mise.local.toml
-ow.toml
-.ow.toml.archived-workspaces
-templates/*
-ow/templates/*
-!ow/templates/common
-!ow/templates/common/**
-!ow/templates/vscode
-!ow/templates/vscode/**
-!ow/templates/zed
-!ow/templates/zed/**
-!ow/templates/bwrap
-!ow/templates/bwrap/**
-docs/superpowers/
 *.egg-info
+.claude/
+.opencode/
+.ow/
 .pytest_cache
 __pycache__
-ow/_version.py
+docs/superpowers/
+src/ow/_version.py
+
+# OW working directories
+.bare-git-repos/
+mise.toml
+mise.local.toml
+ow.toml
+services/*
+templates/*
+workspaces/*
