+++ src/ow/_static/services/compose.yml
