+++ src/ow/_static/templates/common/odoorc.j2
