+++ src/ow/_static/templates/common/pyrightconfig.json.j2
