+++ src/ow/_static/templates/vscode/.vscode/settings.json.j2
