+++ src/ow/_static/templates/zed/.zed/debug.json.j2
