+++ src/ow/_static/templates/zed/.zed/settings.json.j2
