+++ src/ow/commands/__init__.py
@@ -1,0 +1,15 @@
+from ow.commands.create import cmd_create
+from ow.commands.init import cmd_init
+from ow.commands.prune import cmd_prune
+from ow.commands.rebase import cmd_rebase
+from ow.commands.status import cmd_status
+from ow.commands.update import cmd_update
+
+__all__ = [
+    "cmd_create",
+    "cmd_init",
+    "cmd_prune",
+    "cmd_rebase",
+    "cmd_status",
+    "cmd_update",
+]
