+++ src/ow/utils/config.py
@@ -1,11 +1,9 @@
-from __future__ import annotations
-
-import re
 import tomllib
-import tomli_w
 from dataclasses import dataclass, field
 from pathlib import Path
 from typing import Any
+
+import tomli_w
 
 
 @dataclass
